+++ climaf/api.py
@@ -67,7 +67,7 @@ from cmacros   import cmacro
 from driver    import ceval, varOf, cfile, cshow, cMA, cimport, cexport 
 from dataloc   import dataloc 
 from operators import cscript, scripts as cscripts, derive, operators
-from cache     import craz, csync as csave , cdump, cdrop
+from cache     import craz, csync as csave , cdump, cdrop,  clist, cls, crm, cdu, cwc
 from clogging  import clogger, clog, clog_file
 from site_settings import atCNRM, onCiclad