+++ examples/plotmap.py
@@ -15,7 +15,7 @@ ta=time_average(dg)
 ##################################################
 # For reference, here is the declaration of oeprator 'ncview' :
 # cscript('ncview' ,'ncview ${in}' , format=None)
-ncview(ta)
+#ncview(ta)
 
 # A NCL-quality plot, using standard operator plot
 #############################################################